refactor: notice 엔드포인트 리팩토링중
